+++ Trs.IntegerMapper/MapConstants.cs
@@ -5,12 +5,12 @@
         /// <summary>
         /// Represents Null for values and Empty for collections.
         /// </summary>
-        public const uint NullOrEmpty = 0;
+        public const ulong NullOrEmpty = 0;
 
         /// <summary>
         /// All mapped values start from this integer. 
         /// It is here to simplifyu the management of special constants like <see cref="NullOrEmpty"/>
         /// </summary>
-        public const uint FirstMappableInteger = 1;
+        public const ulong FirstMappableInteger = 1;
     }
 }
